requirements.txt
slack-sdk
python-dotenv
sqlalchemy
blake3
schedule
tqdm
openai
llama-index
llama-index-core
llama-index-vector-stores-redis
llama-index-storage-docstore-redis
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
llama-index-postprocessor-flag-embedding-reranker
llama_index.readers.slack
llama_index.embeddings.huggingface
redis
redisvl
colorlog
slack-bolt
pydantic
PyYAML
